requirements.txt
Flask
Flask-Cors
requests
psycopg2
python-dotenv